pyproject.toml
[tool.poetry]
name = "cna_advogado"
version = "0.1.2"
description = ""
authors = ["levyvix <levy.vix@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest-playwright = "^0.5.0"
typer = "^0.12.3"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
selenium = "^4.21.0"


[tool.poetry.scripts]
cna = "cna_advogado.main:app"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
